+++ src/main/java/fr/eitelalaric/gestiondestock/service/impl/CommandeProviderServiceImpl.java
@@ -2,15 +2,19 @@ package fr.eitelalaric.gestiondestock.service.impl;
 
 import fr.eitelalaric.gestiondestock.dto.CommandeProviderDto;
 import fr.eitelalaric.gestiondestock.dto.LigneCommandeProviderDto;
+import fr.eitelalaric.gestiondestock.dto.MvtStkDto;
+import fr.eitelalaric.gestiondestock.dto.ProductDto;
 import fr.eitelalaric.gestiondestock.exception.EntityNotFoundException;
 import fr.eitelalaric.gestiondestock.exception.ErrorCodes;
 import fr.eitelalaric.gestiondestock.exception.InvalidEntityException;
+import fr.eitelalaric.gestiondestock.exception.InvalidOperationException;
 import fr.eitelalaric.gestiondestock.model.*;
 import fr.eitelalaric.gestiondestock.repository.CommandProviderRepository;
 import fr.eitelalaric.gestiondestock.repository.LigneCommandProviderRepository;
 import fr.eitelalaric.gestiondestock.repository.ProductRepository;
 import fr.eitelalaric.gestiondestock.repository.ProviderRepository;
 import fr.eitelalaric.gestiondestock.service.CommandeProviderService;
+import fr.eitelalaric.gestiondestock.service.MvtStkService;
 import fr.eitelalaric.gestiondestock.validator.CommandeProviderValidator;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -25,16 +29,19 @@ import java.util.stream.Collectors;
 @Service
 public class CommandeProviderServiceImpl implements CommandeProviderService {
 
-    private ProviderRepository providerRepository;
-    private ProductRepository productRepository;
-    private CommandProviderRepository commandeProviderRepository;
-    private LigneCommandProviderRepository ligneCommandeProviderRepo;
+    private final ProviderRepository providerRepository;
+    private final ProductRepository productRepository;
+    private final CommandProviderRepository commandeProviderRepository;
+    private final LigneCommandProviderRepository ligneCommandeProviderRepo;
 
-    public CommandeProviderServiceImpl(ProviderRepository providerRepository, ProductRepository productRepository, CommandProviderRepository commandeProviderRepository, LigneCommandProviderRepository ligneCommandeProviderRepo) {
+    private final MvtStkService mvtStkService;
+
+    public CommandeProviderServiceImpl(ProviderRepository providerRepository, ProductRepository productRepository, CommandProviderRepository commandeProviderRepository, LigneCommandProviderRepository ligneCommandeProviderRepo, MvtStkService mvtStkService) {
         this.providerRepository = providerRepository;
         this.productRepository = productRepository;
         this.commandeProviderRepository = commandeProviderRepository;
         this.ligneCommandeProviderRepo = ligneCommandeProviderRepo;
+        this.mvtStkService = mvtStkService;
     }
 
     @Override
@@ -44,9 +51,9 @@ public class CommandeProviderServiceImpl implements CommandeProviderService {
             log.error("Command provider not valid: {}", commandeProviderDto);
             throw new InvalidEntityException("Command provider is not valid", ErrorCodes.COMMAND_PROVIDER_NOT_VALID,errors);
         }
-        Optional<Provider> provider = providerRepository.findById(commandeProviderDto.getProvider().getIdprovider());
+        Optional<Provider> provider = providerRepository.findById(commandeProviderDto.getProvider().getId());
         if (provider.isEmpty()) {
-            log.error("Provider with ID {} is not present in db", commandeProviderDto.getProvider().getIdprovider());
+            log.error("Provider with ID {} is not present in db", commandeProviderDto.getProvider().getId());
             throw new EntityNotFoundException("NO provider with ID found in db", ErrorCodes.PROVIDER_NOT_FOUND);
         }
         List<String> productErrors = new ArrayList<>();
@@ -114,6 +121,25 @@ public class CommandeProviderServiceImpl implements CommandeProviderService {
             log.error("Command provider ID is null");
             return;
         }
+        List<LigneCommandeProvider> commandeProviders =ligneCommandeProviderRepo.findAllByCommandeProviderId(id);
+        if (commandeProviders.isEmpty()) {
+            throw new InvalidOperationException("Impossible de supprimer une commande provider en utilisation",
+                    ErrorCodes.COMMAND_PROVIDER_ALREADY_IN_USE);
+        }
         commandeProviderRepository.deleteById(id);
+    }
+
+    private void updateMvtStk(Integer idCommande) {
+        List<LigneCommandeProvider> ligneCommandeProviders = ligneCommandeProviderRepo.findAllByCommandeProviderId(idCommande);
+        ligneCommandeProviders.forEach(ligneCommandeProvider -> {
+            MvtStkDto mvtStkDto = MvtStkDto.builder()
+                    .product(ProductDto.fromEntity(ligneCommandeProvider.getProduct()))
+                    .sourceMvtStk(SourceMvtStk.COMMANDE_FOURNISSEUR)
+                    .quantite(ligneCommandeProvider.getQuantite())
+                    .idCompany(ligneCommandeProvider.getIdCompany())
+                    .build();
+            mvtStkService.sortieStock(mvtStkDto);
+        });
+
     }
 }
